Remove DICT-GET-ERROR call from DICT-CHECK
enchant_dict_check() returns a negative integer if there was an error.
It's not certain that enchant_dict_get_error() will help describing the
error so don't call it automatically. At least UTF-8 encoding error
messages can be requested with DICT-GET-ERROR function. That function
will be exported in the following commit(s).